requirements.txt
fastapi
uvicorn
python-multipart
numpy
httpx
python-dotenv
Pillow
jinja2
